refactor: move everything to have an include directory
