+++ CMakeLists.txt
@@ -10,6 +10,7 @@ set(CMAKE_C_STANDARD 11)
 set(CMAKE_C_STANDARD_REQUIRED YES)
 set(CMAKE_LIBRARY_PATH lib/)
 set(CMAKE_INCLUDE_PATH include/)
+include_directories(include)
 if (MSVC)
     add_compile_definitions(_CRT_SECURE_NO_WARNINGS)
     # add_compile_options(/Za)
@@ -59,6 +60,7 @@ add_subdirectory(geometry)
 add_subdirectory(graphics)
 add_subdirectory(searching)
 add_subdirectory(conversions)
+message(STATUS panqkg: ${CMAKE_INCLUDE_DIR})
 add_subdirectory(client_server)
 add_subdirectory(project_euler)
 add_subdirectory(machine_learning)
