+++ client_server/CMakeLists.txt
@@ -29,7 +29,7 @@ if(ARPA_HEADERS OR WINSOCK_HEADER)
         endif ()
         if (HAVE_FORK_WINDOWS)
             target_include_directories(${testname} ${CMAKE_INCLUDE_PATH})
-            target_link_libraries(${testname} scilab)
+            target_link_libraries(${testname} forkwindows)
         endif ()
         # if(HAS_UNISTD)
         #     target_compile_definitions(${testname} PRIVATE HAS_UNISTD)
