+++ client_server/tcp_full_duplex_client.c
@@ -27,7 +27,7 @@
 #include <windows.h>
 #include <winsock2.h>
 
-#include "scilab/forkwindows.h"
+#include "../include/scilab/forkwindows.h"
 #define sleep(a) Sleep(a * 1000)
 #else
 #include <arpa/inet.h>  /// For the type in_addr_t and in_port_t
