+++ include/scilab/bool.h
