+++ scilab/forkwindows.c
@@ -13,7 +13,7 @@
  *
  */
 /*--------------------------------------------------------------------------*/
-#include "forkwindows.h"
+#include "../include/scilab/forkwindows.h"
 
 #include <setjmp.h>
 #include <windows.h>
